Fix the solutions

+++ week-05/11/11.cpp
@@ -7,11 +7,11 @@ int main()
 	int n;
 	cin >> n;
 
-	int** table = new int* [n];
+	 double** table = new double* [n];
 
 	for (int i = 0; i < n; i++)
 	{
-		table[i] = new int[n];
+		table[i] = new double[n];
 
 		for (int j = 0; j < n; j++)
 		{
